code cleaned up and organised.
