--- src/TestCases/Exceptions/InvalidParentClassException.php
@@ -1,10 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Eonx\TestUtils\TestCases\Exceptions;
-
-use RuntimeException;
-
-class InvalidParentClassException extends RuntimeException
-{
-}
