+++ tests/Unit/TestCases/AssertArraySameWithDatesTest.php
@@ -8,9 +8,9 @@ use Eonx\TestUtils\TestCases\UnitTestCase;
 use PHPUnit\Framework\AssertionFailedError;
 
 /**
- * @covers \Eonx\TestUtils\TestCases\UnitTestCase
+ * @covers \Eonx\TestUtils\TestCases\UnitTestCase::assertArraySameWithDates
  */
-class UnitTestCaseTest extends UnitTestCase
+class AssertArraySameWithDatesTest extends UnitTestCase
 {
     /**
      * Test array same with dates works fine with date instances.
@@ -93,62 +93,5 @@ class UnitTestCaseTest extends UnitTestCase
         $this->expectExceptionMessage('Failed asserting that two arrays are identical.');
 
         self::assertArraySameWithDates($expected, $actual);
-    }
-
-    /**
-     * Assert that json can be asserted against its decoded array representation.
-     *
-     * @return void
-     *
-     * @throws \Exception
-     */
-    public function testJsonSameAsArray(): void
-    {
-        $json = '{"abc": "xyz", "date": "2019-10-10T01:04:45Z"}';
-        $expected = ['abc' => 'xyz', 'date' => ':fuzzy:'];
-
-        self::assertJsonEqualsStringFuzzily($expected, $json);
-    }
-
-    /**
-     * Assert that json can be asserted against its decoded array representation.
-     * In this case the test fails.
-     *
-     * @return void
-     *
-     * @throws \Exception
-     */
-    public function testJsonSameAsArrayFails(): void
-    {
-        $json = '{"abc": "xyz", "date": "2019-10-10T01:04:45Z"}';
-        $expected = ['xyz' => 'abc', 'date' => ':fuzzy:'];
-
-        $this->expectException(AssertionFailedError::class);
-        $this->expectExceptionMessage('Failed asserting that two arrays are identical.');
-
-        self::assertJsonEqualsStringFuzzily($expected, $json);
-    }
-
-    /**
-     * Assert that json can be asserted against its decoded array representation.
-     * In this case the test fails.
-     *
-     * @return void
-     *
-     * @throws \Exception
-     */
-    public function testJsonSameAsArrayWithInvalidJson(): void
-    {
-        $json = '{"abc": "xyz}';
-        $expected = ['abc' => 'xyz'];
-
-        $this->expectException(AssertionFailedError::class);
-        $this->expectExceptionMessage(
-            <<<EOF
-Failed asserting that '{"abc": "xyz}' is valid JSON (Unexpected control character found).
-EOF
-        );
-
-        self::assertJsonEqualsStringFuzzily($expected, $json);
     }
 }
